move parallelization to separate file
