+++ tinybaker/combinators/merge.py
@@ -2,30 +2,9 @@ from typing import List, Iterable, Any
 from ..transform import Transform, TransformMeta, coerce_to_transform
 from ..exceptions import BakerError, TagConflictError
 from ..util import classproperty
-from threading import Thread
-from multiprocessing import Pool
-from queue import Queue
+from ..parallel import run_parallel
 from typeguard import typechecked
 from .combinatormeta import CombinatorMeta
-
-
-def _mp_run(arg):
-    instance, current_run_info = arg
-    return instance._exec_with_run_info(current_run_info)
-
-
-class MergeWorker(Thread):
-    def __init__(self, queue):
-        Thread.__init__(self)
-        self.queue = queue
-
-    def run(self):
-        while True:
-            instance, run_info = self.queue.get()
-            try:
-                instance._exec_with_run_info(run_info)
-            finally:
-                self.queue.task_done()
 
 
 @typechecked
@@ -109,25 +88,6 @@ def _create_merge_class(merge_steps, merge_input_tags, merge_output_tags, merge_
                         overwrite=self.overwrite,
                     )
                 )
-
-            if self.context.parallel_mode == "multithreading":
-                queue = Queue()
-                for _ in range(min(len(instances), self.context.max_threads)):
-                    worker = MergeWorker(queue)
-                    worker.daemon = True
-                    worker.start()
-
-                for instance in instances:
-                    queue.put((instance, self._current_run_info))
-                queue.join()
-            elif self.context.parallel_mode == "multiprocessing":
-                with Pool(min(len(instances), self.context.max_processes)) as p:
-                    mp_args = [
-                        (instance, self._current_run_info) for instance in instances
-                    ]
-                    p.map(_mp_run, mp_args)
-            else:
-                for instance in instances:
-                    instance._exec_with_run_info(self._current_run_info)
+            run_parallel(instances, self.context, self._current_run_info)
 
     return Merged
